Allow drag-to-scroll *and* drag-and-drop using touch
Adding a delayTouchStart will delay the moment until a
drag and drop action is detected. Before that delay,
if the user does not wait and hold (for longer than 100ms),
and continues to scroll instead, the action will be a
drag-to-scroll, which simply allows the native scroll.

As a consequence: a list of drag sources may intuitively be
scrolled using drag-to-scroll, but once an item is touched
without scrolling for a short moment, the user is able to
drag and drop that element.