Highlight valid dropzones, when an item is dragged
Dropzones will be highlighted, when an item can be dropped in there.
This even works for files that are being dragged.

To do this, I first had to refactor and re-type Dropzone, DropzoneList
and all related components. Now a PossibleDroppableObject type exists,
along all kinds of DragItem* types that make it clearer, which items may
be dropped where.