Add selected item already in state reducer,
Since onStateChange only gets called on state diff,
and selectedItem isn't cleared when an item is removed,
it won't "change", when it's removed and re-added again.
So we didn't get the change.

There also doesn't seem to be a way to clear selectedItem, except for
reset, which would also reset the menu (--> close it).

Solution: we don't use onStateChange anymore, but a side effect of
the stateReducer instead. Doesn't feel great, but seems to work.

The downshift docs are wrong, recommending to call addSelectedItem in
onStateChange, because all of their examples work with a closing menu,
after selecting one item.

We're keeping the menu open.