Decouple distributed execution
This PR introduces a decoupling of the manager and the distributed
execution. This is done mainly by introducing interfaces for the
following classes and moving everything related to remote calls into
a separate implementation:
- `Namespace`
- `DatasetRegistry`
- `AdminDatasetProcessor`

Whenever an existing class requires access to the remote calls,
it either has a hard coupling to the distributed implementation of the
interface or there is an interface or method that can safely result in a
no-op (e.g., `getShardInformation` returns an empty list).