Reworks Printers: We inject the result specific Printers using the printerFactory. This untangles the burden of formatting.
Now a Select claiming to produce a specific type only needs to guarantee they are in a format capable of being handled by the printers it elects to use.

This allows us later to make a printer that alters the result before printing it, without having to inject into all renderers.