create test to reproduce bug
The bug occurs first after a restart, when two concepts connect to the same table and thus to the same buckets. This causes the bucket id to be registered multiple time (the number of connectors). During query time a bucket is then processed multiple times, which renders wrong results in aggregators and aggregation filters.