Rework API, removing PathParam Injection (#3609)
Reworks all API Endpoints that used resolved Id-References on-the-fly. We now explicitly have to resolve all Ids.

This also introduces Id#getDomain, which is the storage or Storage provider of an Id (i.e. the Domain on which it can be resolved). Ids without domains can exist as a kind of query into the store, but this should be avoided. Instead they should be derived from their corresponding object.  
Ids are strictly separated into the MetaStorage domain and the Namespaced domain:

* MetaStorage containing user dependent information or configuration
* Namespaced relates to Datasets and Concept configuration.

NamespacedIds and -Objects try to always delegate to their DatasetId as root for any Domain related tasks. MetaIds and -Objects usually hold their own domain as there are no obvious roots.

---

We refactored the API away from explicit object usage to reduce implicit conversions, but also to make it easier to use with openapi specs. 

Alongside, we also cleaned up the usage of resolve and getId, where they sometimes would be followed by each other in close succession, making them meaningless. We also tried refactoring BucketManager#create, though it still has some circular dependency with WorkerStorage and does IO at inopportune times in startup.